Fix x coordinate calculation in buttonpress.

+++ dwm/dwm.c
@@ -516,7 +516,7 @@ buttonpress(XEvent *e)
 		} else if (ev->x < x + blw)
 			click = ClkLtSymbol;
 		/* 2px right padding */
-		else if (ev->x > selmon->ww - TEXTW(stext) - getsystraywidth() + lrpad - 2)
+		else if (ev->x > selmon->ww - (int)TEXTW(stext) - getsystraywidth() + lrpad - 2)
 			click = ClkStatusText;
 		else {
 			x += blw;
